rectify failing tests
