+++ tests/cli_tests/config/zbox_config.yaml
@@ -1,4 +1,4 @@
-block_worker: https://dev-2.devnet-0chain.net/dns
+block_worker: https://dev-3.devnet-0chain.net/dns
 signature_scheme: bls0chain
 min_submit: 50 # in percentage
 min_confirmation: 50 # in percentage
