+++ tests/cli_tests/flaky___broken_scenarios_test.go
@@ -460,4 +460,102 @@ func Test___FlakyBrokenScenarios(t *testing.T) {
 
 		createAllocationTestTeardown(t, allocationID)
 	})
+
+	t.Run("update blobber read price should work", func(t *testing.T) {
+		if testing.Short() {
+			t.Skip("update blobber read price not working.")
+		}
+		t.Parallel()
+
+		output, err := registerWallet(t, configPath)
+		require.Nil(t, err, "Failed to register wallet", strings.Join(output, "\n"))
+
+		output, err = registerWalletForName(t, configPath, blobberOwnerWallet)
+		require.Nil(t, err, "registering wallet failed", err, strings.Join(output, "\n"))
+
+		output, err = listBlobbers(t, configPath, createParams(map[string]interface{}{"json": ""}))
+		require.Nil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 1, strings.Join(output, "\n"))
+
+		var blobberList []climodel.BlobberDetails
+		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&blobberList)
+		require.Nil(t, err, strings.Join(output, "\n"))
+
+		intialBlobberInfo := blobberList[0]
+
+		oldReadPrice := intialBlobberInfo.Terms.Read_price
+		defer func() {
+			output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "read_price": oldReadPrice}))
+			require.Nil(t, err, strings.Join(output, "\n"))
+			require.Len(t, output, 1)
+		}()
+
+		newReadPrice := oldReadPrice + 1
+
+		// BUG: read price is not being updated
+		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "read_price": newReadPrice}))
+		require.NotNil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 3)
+
+		time.Sleep(30 * time.Second)
+		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
+		require.Nil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 1)
+
+		var finalBlobberInfo climodel.BlobberDetails
+		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
+		require.Nil(t, err, strings.Join(output, "\n"))
+
+		// BUG: read price is not being updated
+		require.NotEqual(t, newReadPrice, finalBlobberInfo.Terms.Read_price)
+	})
+
+	t.Run("update blobber write price should work", func(t *testing.T) {
+		if testing.Short() {
+			t.Skip("update blobber write price not working.")
+		}
+		t.Parallel()
+
+		output, err := registerWallet(t, configPath)
+		require.Nil(t, err, "Failed to register wallet", strings.Join(output, "\n"))
+
+		output, err = registerWalletForName(t, configPath, blobberOwnerWallet)
+		require.Nil(t, err, "registering wallet failed", err, strings.Join(output, "\n"))
+
+		output, err = listBlobbers(t, configPath, createParams(map[string]interface{}{"json": ""}))
+		require.Nil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 1, strings.Join(output, "\n"))
+
+		var blobberList []climodel.BlobberDetails
+		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&blobberList)
+		require.Nil(t, err, strings.Join(output, "\n"))
+
+		intialBlobberInfo := blobberList[0]
+
+		oldWritePrice := intialBlobberInfo.Terms.Write_price
+		defer func() {
+			output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "write_price": oldWritePrice}))
+			require.Nil(t, err, strings.Join(output, "\n"))
+			require.Len(t, output, 1)
+		}()
+
+		newWritePrice := oldWritePrice + 1
+
+		// BUG: write price is not being updated
+		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "write_price": newWritePrice}))
+		require.NotNil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 3)
+
+		time.Sleep(30 * time.Second)
+		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
+		require.Nil(t, err, strings.Join(output, "\n"))
+		require.Len(t, output, 1)
+
+		var finalBlobberInfo climodel.BlobberDetails
+		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
+		require.Nil(t, err, strings.Join(output, "\n"))
+
+		// BUG: write price is not being updated
+		require.NotEqual(t, newWritePrice, finalBlobberInfo.Terms.Write_price)
+	})
 }
