+++ tests/cli_tests/zboxcli_blobber_config_update_test.go
@@ -10,7 +10,7 @@ import (
 
 	climodel "github.com/0chain/system_test/internal/cli/model"
 	cliutils "github.com/0chain/system_test/internal/cli/util"
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/require"
 )
 
@@ -57,6 +57,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -65,8 +66,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
 		require.Nil(t, err, strings.Join(output, "\n"))
 
-		// BUG: capacity is not being updated
-		assert.NotEqual(t, int64(newCapacity), finalBlobberInfo.Capacity)
+		require.Equal(t, int64(newCapacity), finalBlobberInfo.Capacity)
 	})
 
 	t.Run("update blobber challenge completion time should work", func(t *testing.T) {
@@ -101,6 +101,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -109,8 +110,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
 		require.Nil(t, err, strings.Join(output, "\n"))
 
-		// BUG: challenge completion time is not being updated
-		assert.NotEqual(t, newChallengeCompletionTIme, finalBlobberInfo.Terms.Challenge_completion_time)
+		require.Equal(t, newChallengeCompletionTIme, finalBlobberInfo.Terms.Challenge_completion_time)
 	})
 
 	t.Run("update blobber max offer duration should work", func(t *testing.T) {
@@ -145,6 +145,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -153,8 +154,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
 		require.Nil(t, err, strings.Join(output, "\n"))
 
-		// BUG: max offer duration is not being updated
-		assert.Equal(t, newMaxOfferDuration, finalBlobberInfo.Terms.Max_offer_duration)
+		require.Equal(t, newMaxOfferDuration, finalBlobberInfo.Terms.Max_offer_duration)
 	})
 
 	t.Run("update blobber max stake should work", func(t *testing.T) {
@@ -183,12 +183,13 @@ func TestBlobberConfigUpdate(t *testing.T) {
 			require.Len(t, output, 1)
 		}()
 
-		newMaxStake := 10000000000001
+		newMaxStake := oldMaxStake - 1
 
 		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "max_stake": newMaxStake}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -198,7 +199,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 
 		// BUG: max stake is not being updated
-		assert.NotEqual(t, int64(newMaxStake), finalBlobberInfo.StakePoolSettings.MaxStake)
+		require.NotEqual(t, int64(newMaxStake), finalBlobberInfo.StakePoolSettings.MaxStake)
 	})
 
 	t.Run("update blobber min stake should work", func(t *testing.T) {
@@ -227,12 +228,13 @@ func TestBlobberConfigUpdate(t *testing.T) {
 			require.Len(t, output, 1)
 		}()
 
-		newMinStake := 10000000001
+		newMinStake := oldMinStake + 1
 
 		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "min_stake": newMinStake}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -242,7 +244,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 
 		// BUG: min stake is not being updated
-		assert.NotEqual(t, int64(newMinStake), finalBlobberInfo.StakePoolSettings.MinStake)
+		require.NotEqual(t, int64(newMinStake), finalBlobberInfo.StakePoolSettings.MinStake)
 	})
 
 	t.Run("update blobber min lock demand should work", func(t *testing.T) {
@@ -277,6 +279,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -285,8 +288,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
 		require.Nil(t, err, strings.Join(output, "\n"))
 
-		// BUG: min lock demand is not being updated
-		assert.NotEqual(t, newMinLockDemand, finalBlobberInfo.Terms.Min_lock_demand)
+		require.Equal(t, newMinLockDemand, finalBlobberInfo.Terms.Min_lock_demand)
 	})
 
 	t.Run("update blobber number of delegates should work", func(t *testing.T) {
@@ -321,6 +323,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
 
+		time.Sleep(30 * time.Second)
 		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
 		require.Nil(t, err, strings.Join(output, "\n"))
 		require.Len(t, output, 1)
@@ -330,95 +333,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 
 		// BUG: number of delegates is not being updated
-		assert.NotEqual(t, newNumberOfDelegates, finalBlobberInfo.StakePoolSettings.NumDelegates)
-	})
-
-	t.Run("update blobber read price should work", func(t *testing.T) {
-		t.Parallel()
-
-		output, err := registerWallet(t, configPath)
-		require.Nil(t, err, "Failed to register wallet", strings.Join(output, "\n"))
-
-		output, err = registerWalletForName(t, configPath, blobberOwnerWallet)
-		require.Nil(t, err, "registering wallet failed", err, strings.Join(output, "\n"))
-
-		output, err = listBlobbers(t, configPath, createParams(map[string]interface{}{"json": ""}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-
-		var blobberList []climodel.BlobberDetails
-		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&blobberList)
-		require.Nil(t, err, strings.Join(output, "\n"))
-
-		intialBlobberInfo := blobberList[0]
-
-		oldReadPrice := intialBlobberInfo.Terms.Read_price
-		defer func() {
-			output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "read_price": oldReadPrice}))
-			require.Nil(t, err, strings.Join(output, "\n"))
-			require.Len(t, output, 1)
-		}()
-
-		newReadPrice := oldReadPrice + 1
-
-		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "read_price": newReadPrice}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-
-		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-
-		var finalBlobberInfo climodel.BlobberDetails
-		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
-		require.Nil(t, err, strings.Join(output, "\n"))
-
-		// BUG: read price is not being updated
-		assert.NotEqual(t, newReadPrice, finalBlobberInfo.Terms.Read_price)
-	})
-
-	t.Run("update blobber write price should work", func(t *testing.T) {
-		t.Parallel()
-
-		output, err := registerWallet(t, configPath)
-		require.Nil(t, err, "Failed to register wallet", strings.Join(output, "\n"))
-
-		output, err = registerWalletForName(t, configPath, blobberOwnerWallet)
-		require.Nil(t, err, "registering wallet failed", err, strings.Join(output, "\n"))
-
-		output, err = listBlobbers(t, configPath, createParams(map[string]interface{}{"json": ""}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1, strings.Join(output, "\n"))
-
-		var blobberList []climodel.BlobberDetails
-		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&blobberList)
-		require.Nil(t, err, strings.Join(output, "\n"))
-
-		intialBlobberInfo := blobberList[0]
-
-		oldWritePrice := intialBlobberInfo.Terms.Write_price
-		defer func() {
-			output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "write_price": oldWritePrice}))
-			require.Nil(t, err, strings.Join(output, "\n"))
-			require.Len(t, output, 1)
-		}()
-
-		newWritePrice := oldWritePrice + 1
-
-		output, err = updateBlobberInfo(t, configPath, createParams(map[string]interface{}{"blobber_id": intialBlobberInfo.ID, "write_price": newWritePrice}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-
-		output, err = getBlobberInfo(t, configPath, createParams(map[string]interface{}{"json": "", "blobber_id": intialBlobberInfo.ID}))
-		require.Nil(t, err, strings.Join(output, "\n"))
-		require.Len(t, output, 1)
-
-		var finalBlobberInfo climodel.BlobberDetails
-		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
-		require.Nil(t, err, strings.Join(output, "\n"))
-
-		// BUG: write price is not being updated
-		assert.NotEqual(t, newWritePrice, finalBlobberInfo.Terms.Write_price)
+		require.NotEqual(t, newNumberOfDelegates, finalBlobberInfo.StakePoolSettings.NumDelegates)
 	})
 
 	t.Run("update blobber service charge should work", func(t *testing.T) {
@@ -462,7 +377,7 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		require.Nil(t, err, strings.Join(output, "\n"))
 
 		// BUG: service charge is not being updated
-		assert.NotEqual(t, newServiceCharge, finalBlobberInfo.StakePoolSettings.ServiceCharge)
+		require.NotEqual(t, newServiceCharge, finalBlobberInfo.StakePoolSettings.ServiceCharge)
 	})
 
 	// failure scenarios
@@ -525,16 +440,16 @@ func TestBlobberConfigUpdate(t *testing.T) {
 		err = json.NewDecoder(strings.NewReader(output[0])).Decode(&finalBlobberInfo)
 		require.Nil(t, err, strings.Join(output, "\n"))
 
-		assert.NotEqual(t, newWritePrice, finalBlobberInfo.Terms.Write_price)
-		assert.NotEqual(t, newServiceCharge, finalBlobberInfo.StakePoolSettings.ServiceCharge)
-		assert.NotEqual(t, newReadPrice, finalBlobberInfo.Terms.Read_price)
-		assert.NotEqual(t, newNumberOfDelegates, finalBlobberInfo.StakePoolSettings.NumDelegates)
-		assert.NotEqual(t, newMaxOfferDuration, finalBlobberInfo.Terms.Max_offer_duration)
-		assert.NotEqual(t, newCapacity, finalBlobberInfo.Capacity)
-		assert.NotEqual(t, newMinLockDemand, finalBlobberInfo.Terms.Min_lock_demand)
-		assert.NotEqual(t, newMinStake, finalBlobberInfo.StakePoolSettings.MinStake)
-		assert.NotEqual(t, newMaxStake, finalBlobberInfo.StakePoolSettings.MaxStake)
-		assert.NotEqual(t, newChallengeCompletionTIme, finalBlobberInfo.Terms.Challenge_completion_time)
+		require.NotEqual(t, newWritePrice, finalBlobberInfo.Terms.Write_price)
+		require.NotEqual(t, newServiceCharge, finalBlobberInfo.StakePoolSettings.ServiceCharge)
+		require.NotEqual(t, newReadPrice, finalBlobberInfo.Terms.Read_price)
+		require.NotEqual(t, newNumberOfDelegates, finalBlobberInfo.StakePoolSettings.NumDelegates)
+		require.NotEqual(t, newMaxOfferDuration, finalBlobberInfo.Terms.Max_offer_duration)
+		require.NotEqual(t, newCapacity, finalBlobberInfo.Capacity)
+		require.NotEqual(t, newMinLockDemand, finalBlobberInfo.Terms.Min_lock_demand)
+		require.NotEqual(t, newMinStake, finalBlobberInfo.StakePoolSettings.MinStake)
+		require.NotEqual(t, newMaxStake, finalBlobberInfo.StakePoolSettings.MaxStake)
+		require.NotEqual(t, newChallengeCompletionTIme, finalBlobberInfo.Terms.Challenge_completion_time)
 	})
 }
